added a parameter to choose between NavAssistant and naked Nav_2
Other than general flexibility, this is motivated because of a pretty horrible performance overhead in using the nav_assistant make_plan service. Because it is an action-within-a-service situation, it requires async service responses, and that seems to slow things down quite a bit when compared with just spinning until the response comes back (in my tests, a whole order of magnitude slower!)

As a result, (and because several of the GSL methods require calling the make_plan services multiple times in a row before deciding where to move), this had a noticeable impact in the performance of the methods.